backend/requirements.txt
fastapi
uvicorn
joblib
pandas
prophet
python-dotenv
openai>=1.40.0
